+++ common/src/main/java/net/dodogang/plume/Plume.java
@@ -12,9 +12,12 @@ import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 import net.minecraft.world.poi.PointOfInterestType;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public class Plume {
     public static final String MOD_ID = "plume";
+    public static final Logger LOGGER = LogManager.getLogger(MOD_ID);
 
     public static final boolean runDevTests = false;