+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## v0.3.2 - 2021-03-27
 - Add `ToolTags` which allows mods to find if an ItemStack contains a tool.
 - Add `BlockEntityRendererRegistry` to register block entity renderers to block
   entity types.